requirements.txt
azure_storage==0.37.0
Flask==3.0.3
nltk==3.8.1
numpy==1.24.4
tensorflow-cpu==2.15.1

